Split HTTP2JSON into source-sets for mocking (and example)$
