+++ app/src/release/java/com/github/polybooks/utils/HTTP2JSON.kt
@@ -11,7 +11,6 @@ import java.util.concurrent.CompletableFuture
 const val CONNECTION_TIMEOUT = 5000
 const val READ_TIMEOUT = 10000
 
-// TODO this would be the release version and we'd have a separate file also containing the fun url2json() with the mock in debug
 /**
  * Given an URL fetches the json at that address and
  * parses it as JSON
