--- app/src/release/java/com/github/polybooks/utils/exists.kt
@@ -1,1 +1,0 @@
-// TODO delete me as soon as another relevant file is created around there
